+++ src/app/components/common/list/list-item/teach-me-list-item/index.tsx
@@ -8,7 +8,7 @@ import localization from "../../../../../consts/localization";
 import {
   ICourseData,
   CourseState,
-} from "../../../../../interfaces/courses/courses.interface";
+} from "../../../../layout/screens/courses/interface";
 import { IItemComponentProps } from "..";
 
 // components
