+++ src/app/components/layout/screens/courses/course-screen/index.less
