+++ src/app/components/layout/screens/courses/courses-list-screen/index.less
