+++ src/app/components/layout/screens/courses/courses-list-screen/index.tsx
@@ -3,10 +3,10 @@ import React, { useContext } from "react";
 import { IListItem } from "../../../../common/list/list-item";
 import List from "../../../../common/list";
 import TMListItem from "../../../../common/list/list-item/teach-me-list-item";
-import { ICourseData } from "../../courses/interface";
+import { ICourseData } from "../interface";
 import UserDetails from "../../../../common/user/user-details";
 import { TeachMeContext } from "../../../../../App";
-import { parseToCourseListItems } from "../../courses/utils";
+import { parseToCourseListItems } from "../utils";
 
 // styles
 import "./index.less";
