+++ src/app/components/layout/screens/courses/interface.ts
@@ -1,4 +1,4 @@
-import { Icon } from "../../hooks/useIconManager";
+import { Icon } from "../../../../hooks/useIconManager";
 
 export enum CourseState {
   Disabled = "disabled",
