+++ src/app/components/layout/screens/courses/quiz-screen/index.less
