+++ src/app/components/layout/screens/information-screen/interface.ts
