+++ src/app/consts/app.ts
@@ -1,6 +1,5 @@
-import { ICourse } from "../interfaces/courses/courses.interface";
-import { IUserData } from "../interfaces/user/user.interface";
-import { ITMState } from "../interfaces/teachme/teachme.interface";
+import { ICourse } from "../components/layout/screens/courses/interface";
+import { IUserData, ITMState } from "../app.interface";
 
 export const defaultUserData: IUserData = {
   user: {
