+++ src/app/hooks/useIconManager.tsx
@@ -1,5 +1,8 @@
 import React from "react";
-import { CourseState, TaskIcon } from "../interfaces/courses/courses.interface";
+import {
+  CourseState,
+  TaskIcon,
+} from "../components/layout/screens/courses/interface";
 import { iconComponent } from "../consts/icons";
 
 export enum Icon {
