--- src/app/interfaces/user/user.interface.ts
@@ -1,10 +1,0 @@
-interface IUser {
-  firstName: string;
-  LastName: string;
-}
-export interface IUserData {
-  user: IUser;
-  courses: {
-    percentCompletion: number;
-  };
-}
